# Replace Core with Base for better performances