--- a/feather.mli
+++ b/feather.mli
@@ -1,4 +1,4 @@
-open Core
+open Base
 
 type cmd