@charlesetc 2021-05-13
This is great, thanks! And I will probably steal that shebang :)